Preparando vistas para generación de código ATL
