+++ memoria/src/routes/savefiles/+page.svelte
@@ -2,6 +2,7 @@
 <script lang="ts">
     import "../../app.css";
     import { themeStore } from "../../stores";
+    import {downloadXMIFile} from "../../functions/exportdata";
 </script>
 
 <div class="flex flex-col h-full w-full">
@@ -74,6 +75,9 @@
                 'Light'
                     ? 'border-[#855dc7] bg-[#f1e9f9] text-[#855dc7]'
                     : 'border-[#6d44ba] bg-[#231833] text-[#6d44ba]'}"
+                on:click={() => {
+                    downloadXMIFile();
+                }}
             >
                 <span class="my-auto material-symbols-outlined">download</span>
             </button>
